+++ src/style/mod.rs
@@ -29,9 +29,48 @@ use crate::{
     style::height_mode::HeightMode,
     utils::font_ext::FontExt,
 };
+use core::marker::PhantomData;
 use embedded_graphics::{prelude::*, primitives::Rectangle, style::TextStyle};
 
 pub use builder::TextBoxStyleBuilder;
+
+/// Tab size helper
+///
+/// This type makes it more obvious what unit is used to define the width of tabs.
+/// The default tab size is 4 spaces.
+#[derive(Copy, Clone, Debug, Hash, PartialEq, Eq, PartialOrd, Ord)]
+pub struct TabSize<F: Font> {
+    pub(crate) width: i32,
+    _font: PhantomData<F>,
+}
+
+impl<F: Font> Default for TabSize<F> {
+    #[inline]
+    fn default() -> Self {
+        Self::spaces(4)
+    }
+}
+
+impl<F: Font> TabSize<F> {
+    /// Calculate tab size from a number of spaces in the current font.
+    #[inline]
+    pub fn spaces(n: u32) -> Self {
+        let space = F::total_char_width(' ') as i32;
+        // make sure n is at least 1, and the multiplication doesn't overflow
+        let size = (n.max(1) as i32).checked_mul(space).unwrap_or(4 * space);
+
+        Self::pixels(size)
+    }
+
+    /// Define the tab size in pixels.
+    #[inline]
+    pub fn pixels(px: i32) -> Self {
+        Self {
+            width: px,
+            _font: PhantomData,
+        }
+    }
+}
 
 /// Styling options of a [`TextBox`].
 ///
@@ -73,7 +112,7 @@ where
     pub line_spacing: i32,
 
     /// Desired column width for tabs
-    pub tab_size: u32,
+    pub tab_size: TabSize<F>,
 }
 
 impl<C, F, A, V, H> TextBoxStyle<C, F, A, V, H>
@@ -99,7 +138,7 @@ where
             vertical_alignment,
             height_mode,
             line_spacing: 0,
-            tab_size: F::default_tab_size(),
+            tab_size: TabSize::default(),
         }
     }
 
@@ -117,7 +156,7 @@ where
             vertical_alignment,
             height_mode,
             line_spacing: 0,
-            tab_size: F::default_tab_size(),
+            tab_size: TabSize::default(),
         }
     }