+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Unreleased
 
  * `RendererFactory` trait that can be used to create a pixel iterator.
  * Handle tabs `\t` with configurable tab size.
+ * Added `TabSize` struct and related style builder method `tab_size`.
 
 ## Changed: